hal/arm: optimize and unify spinlock implementations
Implement spinlocks as outlined in ARM Architecture Reference Manual
(DDI 0406C.c, sections D7.3.1 and D7.3.2).
Reduce number of registers used for spinlock operations.
Unify implementation across all ARM platforms.

JIRA: RTOS-1304
