+++ hal/armv7a/Makefile
@@ -6,10 +6,13 @@
 
 ifneq (, $(findstring zynq7000, $(TARGET_SUBFAMILY)))
     include hal/armv7a/zynq7000/Makefile
-    CFLAGS += -Ihal/armv7a/zynq7000 -Ihal/armv7a
+    CFLAGS += -Ihal/armv7a/zynq7000
 else ifneq (, $(findstring imx6ull, $(TARGET_SUBFAMILY)))
     include hal/armv7a/imx6ull/Makefile
-    CFLAGS += -Ihal/armv7a/imx6ull -Ihal/armv7a
+    CFLAGS += -Ihal/armv7a/imx6ull
 endif
 
-OBJS += $(addprefix $(PREFIX_O)hal/armv7a/, string.o spinlock.o cpu.o hal.o pmap.o exceptions.o _memcpy.o _memset.o)
+CFLAGS += -Ihal/armv7a -Ihal/arm
+
+OBJS += $(addprefix $(PREFIX_O)hal/arm/, spinlock.o)
+OBJS += $(addprefix $(PREFIX_O)hal/armv7a/, string.o cpu.o hal.o pmap.o exceptions.o _memcpy.o _memset.o)
