+++ hal/armv7m/stm32/config.h
@@ -17,6 +17,8 @@
 #define _PH_HAL_CONFIG_H_
 
 
+#define NUM_CPUS 1
+
 #ifndef __ASSEMBLY__
 #include "hal/types.h"
 #include "include/arch/armv7m/stm32/syspage.h"
