+++ hal/armv8m/Makefile
@@ -7,17 +7,16 @@
 
 ifneq (, $(findstring nrf, $(TARGET_SUBFAMILY)))
     include hal/armv8m/nrf/Makefile
-    CFLAGS += -Ihal/armv8m
 else ifneq (, $(findstring mcx, $(TARGET_SUBFAMILY)))
     include hal/armv8m/mcx/Makefile
-    CFLAGS += -Ihal/armv8m
 else ifneq (, $(findstring stm32, $(TARGET_SUBFAMILY)))
     include hal/armv8m/stm32/Makefile
-    CFLAGS += -Ihal/armv8m
 endif
 
 include hal/arm/Makefile
-CFLAGS += -Ihal/arm
 
-OBJS += $(addprefix $(PREFIX_O)hal/armv8m/, string.o spinlock.o cpu.o hal.o \
+CFLAGS += -Ihal/armv8m -Ihal/arm
+
+OBJS += $(addprefix $(PREFIX_O)hal/arm/, spinlock.o)
+OBJS += $(addprefix $(PREFIX_O)hal/armv8m/, string.o cpu.o hal.o \
   pmap.o exceptions.o _memcpy.o _memset.o interrupts.o)
