+++ hal/armv8m/stm32/n6/config.h
@@ -16,6 +16,7 @@
 #ifndef _PH_HAL_CONFIG_H_
 #define _PH_HAL_CONFIG_H_
 
+#define NUM_CPUS 1
 
 #ifndef __ASSEMBLY__
 #include "hal/types.h"
